Move function to lib
